fix(providers): decide watchdog payload by JSON parse, not first character
The brace-prefix exception was the last gap in a first-character
heuristic: field names may contain any characters per the grammar, so
`{heartbeat}: ping` slipped past the JSON-prefix check and reset the
deadline while every parser discards the line. Classify a line as
payload only when it is a payload-bearing `data:` field at the start of
the line or parses as JSON — exactly what the bare-payload transports
emit — which closes every non-JSON field shape at once. The grammar
classifier stays with the Responses parser, its only remaining
consumer.
