+++ crates/goose-provider-types/src/stream_idle_timeout.rs
@@ -8,21 +8,22 @@
 //! which a stalled stream never produces, and surface a retryable
 //! [`ProviderError::NetworkError`] when the stream goes idle.
 //!
-//! Non-data means: comment frames (`: ...`), blank separators, and every
-//! SSE field line that is not a payload-bearing `data:` field — control
-//! fields (`event:`, `id:`, `retry:`), unknown extension fields
-//! (`x-heartbeat: ...`, `x.heartbeat: ...`, ` data: ...`: the field name
-//! is everything before the first colon, leading whitespace included, so
-//! the parsers that match `data:` at byte zero discard such lines), and
-//! `data:` fields with an empty payload — none of which the downstream
-//! parsers can turn into progress, so none of them may hold the watchdog
-//! off. Lines that are not fields at all because they carry their own
-//! framing (bare JSON frames, Ollama's NDJSON — leading whitespace
-//! notwithstanding) are payloads and do reset the timer.
-//! Payload-bearing events also reset it, including protocol-level pings such
-//! as Anthropic's `data: {"type":"ping"}`: telling those apart from real
-//! progress needs per-provider payload knowledge, which this line-level
-//! watchdog deliberately does not have.
+//! Non-data means: every line that is not a payload-bearing `data:` field
+//! and does not parse as JSON — comment frames (`: ...`), blank
+//! separators, control fields (`event:`, `id:`, `retry:`), `data:` fields
+//! with an empty payload, and unknown extension fields of any shape
+//! (`x-heartbeat: ...`, `x.heartbeat: ...`, `{heartbeat}: ...`,
+//! ` data: ...`: the grammar puts no restriction on field-name characters,
+//! leading whitespace included, and none of these parse as JSON). None of
+//! those can be turned into progress by the downstream parsers, so none
+//! may hold the watchdog off. Lines that parse as JSON — the Responses
+//! API's bare frames and Ollama's NDJSON, leading whitespace
+//! notwithstanding — are payloads and do reset the timer, as do
+//! payload-bearing events including protocol-level pings such as
+//! Anthropic's `data: {"type":"ping"}` and a heartbeat that is itself a
+//! complete JSON object: telling those apart from real progress needs
+//! per-provider payload knowledge, which this line-level watchdog
+//! deliberately does not have.
 //!
 //! The same line-level boundary applies to providers that reassemble a JSON
 //! event split across several lines (Google's parser buffers continuation
@@ -60,46 +61,18 @@ fn resolve_stream_idle_timeout() -> Option<Duration> {
     (secs > 0).then(|| Duration::from_secs(secs))
 }
 
-/// Parse a line per the SSE grammar and return its field name:
-/// - `field: value` / `field:value` -> `Some(field)`
-/// - `field` (no colon, empty value) -> `Some(field)`
-/// - `: comment` -> `Some("")`
-///
-/// The grammar puts no restriction on field-name characters — everything
-/// before the first colon is the name, so `x.heartbeat` is as valid as
-/// `event`, and leading whitespace belongs to the name (` data` is an
-/// unknown field, not `data`). Compliant consumers ignore every field
-/// other than `data`. `None` therefore marks lines that are not fields at
-/// all: payload lines that carry their own framing, i.e. bare JSON frames
-/// and NDJSON — leading whitespace notwithstanding — so callers can treat
-/// them as payload rather than ignorable control fields.
-pub(crate) fn sse_field_name(line: &str) -> Option<&str> {
-    let field = line.split_once(':').map_or(line, |(name, _)| name);
-    if field.is_empty() {
-        return Some("");
-    }
-    let after_whitespace = field.trim_start();
-    let json_payload = after_whitespace.starts_with('{') || after_whitespace.starts_with('[');
-    (!json_payload).then_some(field)
-}
-
 /// Whether a line is progress a downstream parser can consume: a
-/// payload-bearing `data:` field, or a line that is not SSE-field-shaped at
-/// all (bare JSON frames, Ollama's NDJSON). Comment frames, blank
-/// separators, control fields, and unknown extension fields — which per
-/// the grammar include whitespace-prefixed `data:` lines, since the
-/// leading space is part of the field name — are not.
+/// payload-bearing `data:` field at the start of the line, or a line that
+/// parses as JSON — the shapes the bare-payload transports emit (the
+/// Responses API's bare frames, Ollama's NDJSON). Anything else — comment
+/// frames, blank separators, control fields, and unknown extension fields,
+/// whose names per the grammar may contain any characters and leading
+/// whitespace — is not.
 fn is_data_line(line: &str) -> bool {
-    if line.trim().is_empty() {
-        return false;
+    if let Some(value) = line.strip_prefix("data:") {
+        return !value.trim().is_empty();
     }
-    match sse_field_name(line) {
-        Some("data") => line
-            .strip_prefix("data:")
-            .is_some_and(|value| !value.trim().is_empty()),
-        Some(_) => false,
-        None => true,
-    }
+    serde_json::from_str::<serde_json::Value>(line).is_ok()
 }
 
 fn stall_error(idle: Duration, keepalive_frames: u64, env_var: &str) -> anyhow::Error {
@@ -268,6 +241,7 @@ mod tests {
             "x-heartbeat: ping",
             "x.heartbeat: ping",
             " data: ping",
+            "{heartbeat}: ping",
             "{\"chunk\":0}",
             "  {\"chunk\":1}",
             "",
@@ -312,6 +286,8 @@ mod tests {
             "ping",
             " data: ping",
             "   data: {\"beat\":1}",
+            "{heartbeat}: ping",
+            "[heartbeat]: 1",
         ] {
             let err = first_stall_error(with_stream_idle_timeout(
                 data_then_repeating("data: {\"a\":1}", heartbeat),
@@ -427,41 +403,6 @@ mod tests {
         // Blank separators are not keepalive frames; the diagnostic must not
         // claim any arrived.
         assert!(!message.contains("keepalive"), "got: {message}");
-    }
-
-    #[test]
-    fn sse_field_name_classifies_fields_vs_payloads() {
-        // Per the SSE grammar a field name is everything before the first
-        // colon, with no restriction on its characters; only lines that
-        // carry their own framing (bare JSON, NDJSON) are not fields.
-        for (line, field) in [
-            ("data: {\"a\":1}", "data"),
-            ("data:", "data"),
-            ("event: ping", "event"),
-            ("id: 7", "id"),
-            ("retry: 3000", "retry"),
-            (": comment", ""),
-            ("ping", "ping"),
-            ("x-heartbeat: ping", "x-heartbeat"),
-            ("x.heartbeat: ping", "x.heartbeat"),
-            ("x~heartbeat: 1", "x~heartbeat"),
-            ("foo bar: 1", "foo bar"),
-            (" data: ping", " data"),
-            ("\tdata: ping", "\tdata"),
-        ] {
-            assert_eq!(sse_field_name(line), Some(field), "line: {line}");
-        }
-        for payload in [
-            "{\"a\":1}",
-            "{}",
-            "{\"a\":1}\n",
-            "[DONE]",
-            "[1,2]",
-            "  {\"a\":1}",
-            "  [1,2]",
-        ] {
-            assert_eq!(sse_field_name(payload), None, "payload: {payload}");
-        }
     }
 
     #[test]
